requirements.txt
fastmcp>=2.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
requests>=2.31.0
python-dotenv>=1.0.0
uvicorn>=0.20.0
